requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-bigquery==3.14.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
